+++ packages/vendure-plugin-accept-blue/src/api/accept-blue-common-resolvers.ts
@@ -32,6 +32,9 @@ import {
   QueryPreviewAcceptBlueSubscriptionsForProductArgs,
   AcceptBlueCardPaymentMethod,
   MutationUpdateAcceptBlueCheckPaymentMethodArgs,
+  MutationDeleteAcceptBluePaymentMethodArgs,
+  MutationCreateAcceptBlueCardPaymentMethodArgs,
+  MutationCreateAcceptBlueCheckPaymentMethodArgs,
 } from './generated/graphql';
 
 @Resolver()
@@ -187,5 +190,51 @@ export class AcceptBlueCommonResolver {
     { input }: MutationUpdateAcceptBlueCheckPaymentMethodArgs
   ): Promise<GraphqlMutation['updateAcceptBlueCheckPaymentMethod']> {
     return await this.acceptBlueService.updateCheckPaymentMethod(ctx, input);
+  }
+
+  @Mutation()
+  @Allow(Permission.UpdateCustomer, Permission.Authenticated)
+  async deleteAcceptBluePaymentMethod(
+    @Ctx() ctx: RequestContext,
+    @Args() { id }: MutationDeleteAcceptBluePaymentMethodArgs
+  ): Promise<GraphqlMutation['deleteAcceptBluePaymentMethod']> {
+    return await this.acceptBlueService.deletePaymentMethod(ctx, id);
+  }
+
+  @Mutation()
+  @Allow(Permission.UpdateCustomer, Permission.Authenticated)
+  async createAcceptBlueCardPaymentMethod(
+    @Ctx() ctx: RequestContext,
+    @Args() { input, customerId }: MutationCreateAcceptBlueCardPaymentMethodArgs
+  ): Promise<GraphqlMutation['createAcceptBlueCardPaymentMethod']> {
+    if (ctx.apiType === 'admin') {
+      // CustomerId is only defined for admin API
+      return await this.acceptBlueService.createCardPaymentMethod(
+        ctx,
+        input,
+        customerId
+      );
+    }
+    // For Shop API, we use the ctx.activeUserId
+    return await this.acceptBlueService.createCardPaymentMethod(ctx, input);
+  }
+
+  @Mutation()
+  @Allow(Permission.UpdateCustomer, Permission.Authenticated)
+  async createAcceptBlueCheckPaymentMethod(
+    @Ctx() ctx: RequestContext,
+    @Args()
+    { input, customerId }: MutationCreateAcceptBlueCheckPaymentMethodArgs
+  ): Promise<GraphqlMutation['createAcceptBlueCheckPaymentMethod']> {
+    if (ctx.apiType === 'admin') {
+      // CustomerId is only defined for admin API
+      return await this.acceptBlueService.createCheckPaymentMethod(
+        ctx,
+        input,
+        customerId
+      );
+    }
+    // For Shop API, we use the ctx.activeUserId
+    return await this.acceptBlueService.createCheckPaymentMethod(ctx, input);
   }
 }