+++ packages/vendure-plugin-accept-blue/src/api/api-extensions.ts
@@ -214,6 +214,23 @@ const commonApiExtensions = gql`
     sec_code: String
   }
 
+  input CreateAcceptBlueCardPaymentMethodInput {
+    sourceToken: String!
+    expiry_month: Int!
+    expiry_year: Int!
+    avs_address: String
+    avs_zip: String
+    name: String
+  }
+
+  input CreateAcceptBlueCheckPaymentMethodInput {
+    name: String!
+    routing_number: String!
+    account_number: String!
+    account_type: String!
+    sec_code: String!
+  }
+
   extend type Mutation {
     updateAcceptBlueCardPaymentMethod(
       input: UpdateAcceptBlueCardPaymentMethodInput!
@@ -221,11 +238,25 @@ const commonApiExtensions = gql`
     updateAcceptBlueCheckPaymentMethod(
       input: UpdateAcceptBlueCheckPaymentMethodInput!
     ): AcceptBlueCheckPaymentMethod!
+    deleteAcceptBluePaymentMethod(id: Int!): Boolean!
   }
 `;
 
 export const shopApiExtensions = gql`
   ${commonApiExtensions}
+
+  extend type Mutation {
+    """
+    Creating a card payment method is only allowed with a nonce token.
+    See Hosted Tokenization in the Accept Blue docs for more information.
+    """
+    createAcceptBlueCardPaymentMethod(
+      input: CreateAcceptBlueCardPaymentMethodInput!
+    ): AcceptBlueCardPaymentMethod!
+    createAcceptBlueCheckPaymentMethod(
+      input: CreateAcceptBlueCheckPaymentMethodInput!
+    ): AcceptBlueCheckPaymentMethod!
+  }
 `;
 
 export const adminApiExtensions = gql`
@@ -247,5 +278,18 @@ export const adminApiExtensions = gql`
     updateAcceptBlueSubscription(
       input: UpdateAcceptBlueSubscriptionInput!
     ): AcceptBlueSubscription!
+
+    """
+    Creating a card payment method is only allowed with a nonce token.
+    See Hosted Tokenization in the Accept Blue docs for more information.
+    """
+    createAcceptBlueCardPaymentMethod(
+      input: CreateAcceptBlueCardPaymentMethodInput!
+      customerId: Int!
+    ): AcceptBlueCardPaymentMethod!
+    createAcceptBlueCheckPaymentMethod(
+      input: CreateAcceptBlueCheckPaymentMethodInput!
+      customerId: Int!
+    ): AcceptBlueCheckPaymentMethod!
   }
 `;